SEEMS TO BE WORKING HUH

+++ disk.h
@@ -1,15 +1,15 @@
 #pragma once
 
 typedef struct fakedisk {
-    unsigned int nblocks;
+    unsigned long nblocks;
     unsigned int blocksize;
     char data[0];
 } disk_t;
 
-disk_t *disk_create(int nblocks, int blocksize);
+disk_t *disk_create(unsigned long nblocks, int blocksize);
 
 void disk_destroy(disk_t *disk);
 
-void disk_read(disk_t *disk, int blockno, void* block);
+void disk_read(disk_t *disk, unsigned long blockno, void* block);
 
-void disk_write(disk_t *disk, int blockno, void* block);
+void disk_write(disk_t *disk, unsigned long blockno, void* block);
